pyproject.toml
[project]
name = "115a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==4.2.7",
    "django-cors-headers==4.2.0",
    "djangorestframework==3.14.0",
    "python-decouple==3.8",
]
